Measure accuracy
Measuring a "missed" stroke is surprisingly hard.  It's actually
impossible to get a perfect reading of whether a word was a misstroke
because of how Plover works.  Fortunately, the process doesn't need to
be 100% accurate.  It's more important that users have a relative
sense for their progress and that's doable here.

The accuracy process works off a signal sent by the input line edit.
The signal passes the contents on change.

Plover works by translating strokes into phrases.  These phrases are
input by Plover one character at a time.  For multi-stroke phrases,
the second stroke may undo what was written by the first; sometimes it
deletes everything!  For example, the strokes "HREBG/TOR" first writes
"lecture" and then, after deleting everything, writes "elector".  This
means we can't simply look for the first non-matching character.  Can
we check for what stroke was entered and see if that matches the
dictionary?  Unfortunately, we only have the resulting phrase.  To
make it worse, there are often multiple strokes corresponding to a
phrase.

I see only three possible dimensions in which to compare the two
strings: spelling, length, and interval between entry.  These give
hints into how we might detect a mistake.

Here are some things to think about.  A misstroke has an incorrect
character.  However, an incorrect character doesn't mean it's a
misstroke (e.g. "HREBG/TOR"->"lecture"->"elector").  An input length
which exceeds the target's length may be a mistake unless a
multi-stroke phrase is being entered.  Generally, a multi-stroke
phrase preserves the first character.  It makes intuitive sense and
also checks out to be a reasonable assumption.  It's also true that a
misstroke must have a return to the original position. Unfortunately,
a multi-stroke phrase may also return to the original position.
Finally, it's reasonable to assume that the user inputs slower than
Plover.

The final assumption, that a user is limited asymptotically by the
speed Plover modifies text, gives us a way to measure accuracy.  We
set checks for whether an entry is "maybe a miss":

- Is there a non-matching character?
- Is the content longer than the target?
- Has the input returned to the starting position after first
  entering text?

We also check to see how much time elapsed between character input.
In testing, Plover entered text 10-100 times faster than a
human (i.e. me).  That's enough of a threshold to make timing a viable
approach[1].  If the time between two chars is 10x slower than the
previous and there is suspicion that a mistake has been made, it's
likely a mistake has been made.

When Plover deletes the word "lecture" within the strokes for
"elector", it does not pause longer between one character or
another. It consistently enters characters on the order of 0.001
seconds between characters.  Assuming 5 characters per word (as
Wolfram Alpha says is true of English), then

x words/minute = (5*x char)/minute * 1 minute/60 seconds

or

x words/minute = x/12 characters/second

We want how many seconds per character.  So,

12/(x words/minute) = y seconds/character

Thus, 12/30wpm => 0.4s between characters. That's at least 10x longer
than Plover's 0.001s between characters.  This is the basis for
TIME_BETWEEN_STROKES.

Finally, if we only allow one mistake per unit/phrase, the accuracy
gives us what words the user was able to enter without error. This is
a useful measure.

[1] I suspect there are plenty of things which can interfere with the
timing.  However, with the 10x-100x margin, it may be crazy enough to
work.  The worst outcome appears to be telling someone they're less
accurate than they really are.