Fix About window not showing
When switching the application to run as a module, the icon could no
longer be accessed directly.  Instead, it was read in as bytes and
converted to a QIcon.  The code for the About window still referenced
the relative path. This caused the window to be created yet not
render.  Labels can't have embedded QPixmaps or QIcons in them.  They
can only have a QPixmap set or embed an image by path.  So, created a
separate label to display the large icon.