Fix inconsistent terminology
Instead of "word" or "phrase", use "unit".  The dictionary maps a
stroke to some unit, such as a word or phrase.  However, since the
output can really be more than a word or phrase (e.g. symbols,
puntuation, partial phrases, control characters, nonsense, etc.), the
aim was to find a more precise term.  It's probably the case that in
search of something "more precise", an awkward term has been
introduced.  At least the dictionary value now has a consistent name.